fix warnings

+++ Parser/Gmsh.yy.cpp
@@ -1039,7 +1039,7 @@ void   skipline(void);
 #define YY_INPUT(buf,result,max_size)                                   \
      {                                                                  \
        int c = '*', n;                                                  \
-       for ( n = 0; n < max_size &&                                     \
+       for ( n = 0; n < (int) max_size &&                               \
                (c = gzgetc( gmsh_yyin )) != EOF && c != '\n'; ++n )          \
          buf[n] = (char) c;                                             \
        if ( c == '\n' ){                                                \
@@ -1058,7 +1058,7 @@ void   skipline(void);
 #define YY_INPUT(buf,result,max_size)					\
      {									\
        int c = '*', n;							\
-       for ( n = 0; n < max_size &&					\
+       for ( n = 0; n < (int) max_size &&                               \
 	       (c = fgetc( gmsh_yyin )) != EOF && c != '\n'; ++n )		\
 	 buf[n] = (char) c;						\
        if ( c == '\n' ){						\
